requirements.txt
streamlit>=1.45.1
openai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.13
google-generativeai>=0.3.0
chromadb>=0.4.18
pypdf>=4.0.0
numpy>=1.24.0
